+++ templates/app/src/app/app.js
@@ -1,5 +1,5 @@
 import angular from 'angular';
-<% if (ngAnimate) { -%>
+<% if (ngAnimate || uiFramework === 'bootstrap') { -%>
 import ngAnimate from 'angular-animate';
 <% } -%>
 import ngAria from 'angular-aria';
@@ -12,8 +12,14 @@ import ngMessages from 'angular-messages';
 <% if (ngSanitize) { -%>
 import ngSanitize from 'angular-sanitize';
 <% } -%>
-<% if (ngTouch) { -%>
+<% if ((ngTouch || uiFramework === 'bootstrap') && uiFramework !== 'ngMaterial') { -%>
 import ngTouch from 'angular-touch';
+<% } -%>
+<% if (uiFramework === 'ngMaterial') { -%>
+import ngMaterial from 'angular-material';
+<% } -%>
+<% if (uiFramework === 'bootstrap') { -%>
+import ngBootstrap from 'angular-ui-bootstrap';
 <% } -%>
 <% if (angularTranslate) { -%>
 import ngTranslate from 'angular-translate';
@@ -35,6 +41,12 @@ import '../styles/main.less';
 <% } else { -%>
 import '../styles/main.css';
 <% } -%>
+<% if (uiFramework === 'ngMaterial') { -%>
+import 'angular-material/angular-material.css';
+<% } -%>
+<% if (uiFramework === 'bootstrap') { -%>
+import 'bootstrap/dist/css/bootstrap.min.css';
+<% } -%>
 
 import config from 'app.config';
 
@@ -43,7 +55,7 @@ import appRoute from './app.route';
 import appComponent from './app.component';
 
 angular.module('app', [
-<% if (ngAnimate) { -%>
+<% if (ngAnimate || uiFramework === 'bootstrap') { -%>
   ngAnimate,
 <% } -%>
   ngAria,
@@ -56,8 +68,14 @@ angular.module('app', [
 <% if (ngSanitize) { -%>
   ngSanitize,
 <% } -%>
-<% if (ngTouch) { -%>
+<% if ((ngTouch || uiFramework === 'bootstrap') && uiFramework !== 'ngMaterial') { -%>
   ngTouch,
+<% } -%>
+<% if (uiFramework === 'ngMaterial') { -%>
+  ngMaterial,
+<% } -%>
+<% if (uiFramework === 'bootstrap') { -%>
+  ngBootstrap,
 <% } -%>
 <% if (angularTranslate) { -%>
   ngTranslate,