+++ templates/app/webpack.config.js
@@ -31,6 +31,10 @@ let common = {
       loader: 'raw'
     },
     {
+      test: /\.css$/,
+      loader: 'style-loader!css-loader'
+    },
+    {
       test: /\.json$/,
       loader: 'json'
     },